refactor(core): split component/inventory god-file into module dir
