+++ src/core/component/inventory/mod.rs
@@ -1,0 +1,12 @@
+mod listing;
+mod path_diagnostics;
+mod reconcile;
+mod types;
+
+pub use listing::*;
+pub use path_diagnostics::*;
+pub use reconcile::*;
+pub use types::*;
+
+#[cfg(test)]
+mod tests;
